backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
pydantic>=2
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
greenlet
python-multipart